correccion y optimizacion de los componentes productDeskyopView, productPhoneView y products
