+++ src/components/Products.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import Product from "./Product";
 import { assets } from "../assets.js";
 
 const products = [
@@ -115,19 +114,6 @@ const products = [
   },
 ];
 
-const ProductList = () => {
-  return (
-    <section id="product" className="bg-gray-200 py-6 px-6 hidden md:block">
-      <div className="text-center py-3">
-        <h2 className="text-4xl font-bold py-8 mb-8">Algunos de Nuestros Productos</h2>
-      </div>
-      <div className="grid grid-cols-1 lg:grid-cols-2 gap-3">
-        {products.map((product, index) => (
-          <Product key={index} {...product} />
-        ))}
-      </div>
-    </section>
-  );
-};
 
-export default ProductList;
+
+export default products;
